requirements.txt
# Telegram
python-telegram-bot==21.7

# Локальная LLM (клиент к Ollama)
ollama>=0.4.0

# Переменные окружения
python-dotenv==1.0.1
